simple_lmk: Make reclaim deterministic
The 20 ms delay in the reclaim thread is a hacky fudge factor that can
cause Simple LMK to behave wildly differently depending on the
circumstances of when it is invoked. When kswapd doesn't get enough CPU
time to finish up and go back to sleep within 20 ms, Simple LMK performs
superfluous reclaims.

This is suboptimal, so make Simple LMK more deterministic by eliminating
the delay and instead queuing up reclaim requests from kswapd.

Change-Id: Iae78315435d1731a700c1ecccc669d068b7867ab
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Hanif Ardhani <hanif.ardhani98@gmail.com>