simple_lmk: Be extra paranoid if tasks can have no pages
If it's possible for a task to have no pages, then there could be a case
where `pages_found` is zero while `nr_found` isn't, which would cause
the found tasks' locks to never be unlocked, and thus mayhem. We can
change the `pages_found` check to use `nr_found` instead in order to
naturally defend against this scenario, in case it is indeed possible.

Change-Id: I200fc6181c6cb25c51f99cafa566735bf74cfb32
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Hanif Ardhani <hanif.ardhani98@gmail.com>